fix memory leaks

See merge request graphviz/graphviz!3948

+++ cmd/smyrna/gltemplate.c
@@ -116,7 +116,7 @@ static gboolean configure_event(GtkWidget *widget, GdkEventConfigure *event,
     glMatrixMode(GL_PROJECTION);
     glLoadIdentity();
     if (widget->allocation.width > 1)
-	init_arcBall(view->arcball, (float)view->w, (float)view->h);
+	init_arcBall(view->arcball, view->w, view->h);
 
     if (view->w > view->h) {
 	aspect = (float) view->w / (float) view->h;
